chore: api for user & organize user apis
