+++ lib/types/fetcher.ts
@@ -4,6 +4,18 @@ import * as schema from '@app/drizzle/schema';
 import * as convert from './convert.ts';
 import type * as res from './res.ts';
 
+export async function fetchSlimUserByUsername(username: string): Promise<res.ISlimUser | null> {
+  const data = await db
+    .select()
+    .from(schema.chiiUser)
+    .where(op.eq(schema.chiiUser.username, username))
+    .execute();
+  for (const d of data) {
+    return convert.toSlimUser(d);
+  }
+  return null;
+}
+
 export async function fetchSlimSubjectByID(
   id: number,
   allowNsfw = false,
