+++ routes/private/routes/login.ts
@@ -37,7 +37,7 @@ const clientPermission = t.Object(
   { $id: 'Permission' },
 );
 
-const currentUser = t.Intersect([res.User, t.Object({ permission: clientPermission })], {
+const currentUser = t.Intersect([res.SlimUser, t.Object({ permission: clientPermission })], {
   $id: 'CurrentUser',
 });
 
@@ -154,7 +154,7 @@ dev.bgm38.com 域名使用测试用的 site-key \`1x00000000000000000000AA\``,
         operationId: 'login',
         tags: [Tag.User],
         response: {
-          200: t.Ref(res.User, {
+          200: t.Ref(res.SlimUser, {
             headers: {
               'Set-Cookie': t.String({ description: `example: "${session.CookieKey}=12345abc"` }),
             },
@@ -185,7 +185,7 @@ dev.bgm38.com 域名使用测试用的 site-key \`1x00000000000000000000AA\``,
     async function loginHandler(
       { body: { email, password, 'cf-turnstile-response': cfCaptchaResponse }, ip },
       reply,
-    ): Promise<res.IUser> {
+    ): Promise<res.ISlimUser> {
       const limitKey = `${redisPrefix}-login-rate-limit-${ip}`;
       const { remain, reset, limit, limited } = await limiter.get(limitKey, 600, 10);
       void reply.headers({
@@ -220,8 +220,8 @@ dev.bgm38.com 域名使用测试用的 site-key \`1x00000000000000000000AA\``,
 
       return {
         ...user,
-        user_group: user.groupid,
         avatar: avatar(user.avatar),
+        joinedAt: user.regdate,
       };
     },
   );
