+++ routes/private/routes/post.test.ts
@@ -64,9 +64,8 @@ describe('get group post', () => {
             "small": "https://lain.bgm.tv/pic/user/s/icon.jpg",
           },
           "id": 287622,
+          "joinedAt": 0,
           "nickname": "nickname 287622",
-          "sign": "sing 287622",
-          "user_group": 0,
           "username": "287622",
         },
         "id": 2092074,
@@ -110,9 +109,8 @@ describe('get subject post', () => {
             "small": "https://lain.bgm.tv/pic/user/s/icon.jpg",
           },
           "id": 2,
+          "joinedAt": 0,
           "nickname": "nickname 2",
-          "sign": "sing 2",
-          "user_group": 11,
           "username": "2",
         },
         "id": 3,
@@ -358,9 +356,8 @@ describe('create group post reply', () => {
           small: 'https://lain.bgm.tv/pic/user/s/icon.jpg',
         },
         id: 1,
+        joinedAt: 0,
         nickname: 'nickname 1',
-        sign: 'sing 1',
-        user_group: 0,
         username: '1',
       },
       state: 0,
